EVA_backend/pyproject.toml
[tool.poetry]
name = "eva"
version = "0.1.0"
description = "AI Agent"
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "memory_service_app" },
    { include = "llm_manager_app" },
    { include = "api_service" },
    { include = "eva_service" },
    { include = "task_manager_app" },
    { include = "eva_backend_django" }
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiofiles = "24.1.0"
aiohappyeyeballs = "2.4.3"
aiohttp = "^3.10.0"
aiolimiter = "1.2.1"
aioredis = "2.0.1"
aiosignal = "1.3.2"
altair = "5.5.0"
amqp = "5.3.1"
annotated-types = "0.7.0"
anyio = "3.6.2"
asgiref = "3.8.1"
asttokens = "2.4.0"
asyncio = "3.4.3"
attrs = "24.2.0"
autobahn = "24.4.2"
autoflake = "2.3.1"
automat = "24.8.1"
av = "11.0.0"
backcall = "0.2.0"
beautifulsoup4 = "4.12.3"
bidict = "0.23.1"
billiard = "4.2.1"
black = "24.10.0"
bleach = "6.2.0"
blinker = "1.8.2"
build = "1.2.2.post1"
celery = "5.4.0"
certifi = "2024.8.30"
cffi = "1.17.1"
cfgv = "3.4.0"
channels = "4.2.0"
channels-redis = "4.0.0"
charset-normalizer = "3.4.0"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
comtypes = "1.4.7"
constantly = "23.10.4"
contourpy = "1.3.1"
cron-descriptor = "1.4.5"
cryptography = "43.0.3"
cycler = "0.12.1"
daphne = "4.1.2"
dataclasses-json = "0.6.7"
decorator = "5.1.1"
defusedxml = "0.7.1"
diskcache = "5.6.3"
distlib = "0.3.9"
distro = "1.9.0"
dj-database-url = "1.0.0"
django = "5.1.4"
django-celery-beat = "2.7.0"
django-cors-headers = "4.5.0"
django-debug-toolbar = "4.4.6"
django-environ = "0.11.2"
django-health-check = "3.18.3"
django-timezone-field = "7.0"
djangorestframework = "3.14.0"
dnspython = "2.7.0"
docopt = "0.6.2"
edge-tts = "7.0.0"
einops = "0.8.0"
encodec = "0.1.1"
eventlet = "0.33.3"
executing = "2.1.0"
fastapi = "0.115.6"
fastjsonschema = "2.21.1"
ffmpy = "0.5.0"
filelock = "3.16.1"
flake8 = "7.1.1"
flake8-bugbear = "24.10.31"
flake8-docstrings = "1.7.0"
fonttools = "4.55.3"
frozenlist = "1.4.1"
fsspec = "2024.6.1"
geopy = "2.4.1"
gevent = "24.2.1"
gevent-websocket = "0.10.1"
googleapis-common-protos = "1.65.0"
greenlet = "3.1.1"
grpcio = "1.66.2"
gunicorn = "23.0.0"
h11 = "0.14.0"
httpcore = "1.0.6"
httpx = "0.27.2"
httpx-sse = "0.4.0"
huggingface-hub = "0.25.1"
hyperlink = "21.0.0"
identify = "2.6.1"
idna = "3.10"
importlib-resources = "6.5.2"
incremental = "24.7.2"
iniconfig = "2.0.0"
ipython = "8.12.3"
isort = "5.13.2"
itsdangerous = "2.2.0"
jedi = "0.19.2"
jinja2 = "3.1.4"
jiter = "0.8.2"
joblib = "1.4.2"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jupyter-client = "8.6.3"
jupyter-core = "5.7.2"
jupyterlab-pygments = "0.3.0"
kiwisolver = "1.4.8"
kombu = "5.4.2"
langchain = "0.3.8"
langchain-community = "0.3.8"
langchain-core = "^0.3.34"
langchain-huggingface = "0.1.2"
langchain-openai = "0.3.0"
langchain-text-splitters = "0.3.2"
langid = "1.1.6"
langsmith = "0.1.129"
llvmlite = "0.43.0"
loguru = "0.7.2"
lz4 = "4.3.3"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.22.0"
matplotlib = "3.10.0"
matplotlib-inline = "0.1.7"
mccabe = "0.7.0"
mdurl = "0.1.2"
mistune = "3.1.0"
mpmath = "1.3.0"
msgpack = "1.1.0"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
narwhals = "1.21.1"
nbclient = "0.10.2"
nbconvert = "7.16.5"
nbformat = "5.10.4"
networkx = "3.3"
nltk = "3.8.1"
nodeenv = "1.9.1"
numba = "0.60.0"
numpy = "1.26.4"
openai = "1.59.6"
opencv-python = "4.10.0.84"
orjson = "3.10.7"
packaging = "24.1"
pandas = "2.2.3"
pandocfilters = "1.5.1"
parso = "0.8.4"
pathspec = "0.12.1"
pexpect = "4.9.0"
pickleshare = "0.7.5"
pillow = "10.4.0"
pip-autoremove = "0.10.0"
pip-check-reqs = "2.5.3"
pip-tools = "7.4.1"
pipdeptree = "2.24.0"
pipreqs = { version = "0.5.0", python = "<3.13" }
pluggy = "1.5.0"
pre-commit = "4.0.1"
prompt-toolkit = "3.0.48"
protobuf = ">=3.20.0,<3.21.0"
psutil = "6.1.0"
psycopg2 = "^2.9.10"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pybase16384 = "0.3.7"
pycodestyle = "2.12.1"
pycparser = "2.22"
pydantic = "2.9.2"
pydantic-settings = "2.5.2"
pydantic-core = "2.23.4"
pydocstyle = "6.3.0"
pydub = "0.25.1"
pyflakes = "3.2.0"
pygments = "2.19.1"
pyopenssl = "24.2.1"
pyparsing = "3.2.1"
pyproject-hooks = "1.2.0"
pytesseract = "0.3.13"
pytest-django = "4.9.0"
pytest-mock = "3.14.0"
python-crontab = "3.2.0"
python-dateutil = "2.9.0.post0"
python-decouple = "3.8"
python-dotenv = "1.0.1"
python-engineio = "4.9.1"
python-multipart = "0.0.20"
python-socketio = "5.9.0"
pytz = "2024.2"
pyyaml = "6.0.2"
pyzmq = "26.2.0"
redis = "5.2.1"
referencing = "0.35.1"
regex = "2024.9.11"
requests = "2.32.3"
retrying = "1.3.4"
rich = "13.9.4"
rpds-py = "0.22.3"
ruff = "0.8.1"
safehttpx = "0.1.6"
safetensors = "0.4.5"
scikit-learn = "1.6.1"
scipy = "1.12.0"
semantic-version = "2.10.0"
sentence-transformers = "3.1.1"
sentry-sdk = "2.17.0"
service-identity = "24.2.0"
shellingham = "1.5.4"
simple-websocket = "1.0.0"
six = "1.16.0"
sniffio = "1.3.1"
snowballstemmer = "2.2.0"
soupsieve = "2.6"
speechrecognition = "3.10.4"
sqlalchemy = "2.0.35"
sqlparse = "0.5.1"
srt = "3.5.3"
stack-data = "0.6.3"
starlette = "0.41.3"
sympy = "1.13.3"
tabulate = "0.9.0"
tenacity = "8.5.0"
textblob = "0.18.0.post0"
threadpoolctl = "3.5.0"
tiktoken = "^0.7.0,<0.8.0"
tinycss2 = "1.4.0"
tokenizers = "0.20.0"
tomlkit = "0.13.2"
tornado = "6.4.2"
tqdm = "4.66.5"
transformers = "4.45.1"
twisted = "24.10.0"
txaio = "23.1.1"
typer = "0.15.1"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
uvicorn = "0.34.0"
vector-quantize-pytorch = "1.12.4"
vine = "5.1.0"
virtualenv = "20.27.1"
vocos = "0.1.0"
watchdog = "5.0.3"
wcwidth = "0.2.13"
webencodings = "0.5.1"
werkzeug = "3.1.1"
wsproto = "1.2.0"
yarg = "0.1.9"
yarl = "1.13.1"
zope-event = "5.0"
zope-interface = "7.0.3"
whitenoise = "^6.6.0"
torch = "^2.7.0"
websockets = "^11.0"
pinecone-client = "2.2.3"
flask = "^3.0.0"
graphrag = "^0.1.0"
datasets = "^2.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
pytest-asyncio = "0.23.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
